+++ lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:app_wisata_flutter/detail_screen.dart';
 
 void main() {
   runApp(const MyApp());
@@ -13,71 +14,6 @@ class MyApp extends StatelessWidget {
       title: 'Wisata Bandung',
       theme: ThemeData(),
       home: const DetailScreen(),
-    );
-  }
-}
-
-class DetailScreen extends StatelessWidget {
-  const DetailScreen({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: SafeArea(
-        child: Column(
-          //Mengubah Ukuran Lebar Layar Sesuai dengan Layar
-          crossAxisAlignment: CrossAxisAlignment.stretch,
-          children: <Widget>[
-          Container(
-            margin: const EdgeInsets.only(top: 16.0),
-            child: const Text(
-              'Farm House Lembang',
-              textAlign: TextAlign.center,
-              style: TextStyle(
-                fontSize: 30.0,
-                fontWeight: FontWeight.bold,
-              ),
-              ),
-          ),
-          Container(
-            margin: const EdgeInsets.symmetric(vertical: 16.0),
-            child: const Row(
-              mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-              children: <Widget>[
-                Column(
-                  children: <Widget>[
-                    Icon(Icons.calendar_today),
-                    SizedBox(height: 8.0),
-                    Text('Open Everyday')
-                  ],
-                ),
-                Column(
-                  children: <Widget>[
-                    Icon(Icons.access_time),
-                    SizedBox(height: 8.0),
-                    Text('09:00 - 20:00')
-                  ],
-                ),
-                Column(
-                  children: <Widget>[
-                    Icon(Icons.monetization_on),
-                    SizedBox(height: 8.0),
-                    Text('RP. 25.000')
-                  ],
-                ),
-              ],
-            )
-          ),
-          Container(
-            padding: const EdgeInsets.all(16.0),
-            child: const Text(
-              'Berada di jalur utama Bandung-Lembang, Farm House menjadi objek wisata yang tidak pernah sepi pengunjung. Selain karena letaknya strategis, kawasan ini juga menghadirkan nuansa wisata khas Eropa. Semua itu diterapkan dalam bentuk spot swafoto Instagramable.',
-              textAlign: TextAlign.center,
-              style: TextStyle(fontSize: 16.0),
-            ),
-          ),
-        ]),
-      )
     );
   }
 }
